Add cluster configuration and mode support

+++ cloud/src/main/java/de/redstonecloud/cloud/RedstoneCloud.java
@@ -5,6 +5,8 @@ import de.redstonecloud.api.encryption.cache.KeyCache;
 import de.redstonecloud.api.redis.broker.BrokerHelper;
 import de.redstonecloud.api.util.Keys;
 import de.redstonecloud.cloud.config.CloudConfig;
+import de.redstonecloud.cloud.config.entry.ClusterEntry;
+import de.redstonecloud.cloud.config.entry.ClusterMode;
 import de.redstonecloud.cloud.config.entry.RedisEntry;
 import de.redstonecloud.cloud.events.EventManager;
 import de.redstonecloud.cloud.player.PlayerManager;
@@ -42,6 +44,8 @@ public class RedstoneCloud {
     private static RedisInstance redisInstance;
     @Getter
     private static boolean running = false;
+    @Getter
+    private static ClusterEntry clusterConfig;
 
 
     @Getter private static Broker broker;
@@ -53,12 +57,14 @@ public class RedstoneCloud {
         if (!Directories.setupCheck.exists()) Utils.setup();
 
         RedisEntry redisCfg = CloudConfig.getRedis();
+        clusterConfig = CloudConfig.getCluster();
 
         System.setProperty(Keys.PROPERTY_REDIS_PORT, redisCfg.port());
         System.setProperty(Keys.PROPERTY_REDIS_IP, redisCfg.ip());
         System.setProperty(Keys.PROPERTY_REDIS_DB, String.valueOf(redisCfg.db()));
 
-        if(redisCfg.useInternal()) {
+        // Only start internal Redis if not using custom Redis AND (standalone mode OR master mode)
+        if(redisCfg.useInternal() && (!clusterConfig.enabled() || clusterConfig.mode() == ClusterMode.MASTER)) {
             redisInstance = new RedisInstance();
         }
 
@@ -117,6 +123,13 @@ public class RedstoneCloud {
         this.keyCache.addKey("cloud", publicKey);
 
         log.info(Translator.translate("cloud.startup"));
+        
+        // Log cluster information
+        if (clusterConfig.enabled()) {
+            log.info("Cluster mode enabled: {} (Node ID: {})", clusterConfig.mode(), clusterConfig.nodeId());
+        } else {
+            log.info("Running in standalone mode");
+        }
 
         Utils.createBaseFolders();
 
@@ -134,7 +147,10 @@ public class RedstoneCloud {
         this.consoleThread = new ConsoleThread();
         this.consoleThread.start();
 
-        this.scheduler.scheduleRepeatingTask(new CheckTemplateTask(), 3000L);
+        // Only schedule template checking task for STANDALONE or MASTER modes
+        if (!clusterConfig.enabled() || clusterConfig.mode() == ClusterMode.MASTER || clusterConfig.mode() == ClusterMode.STANDALONE) {
+            this.scheduler.scheduleRepeatingTask(new CheckTemplateTask(), 3000L);
+        }
 
         this.pluginManager.enableAllPlugins();
     }
@@ -151,8 +167,13 @@ public class RedstoneCloud {
         try {
             Thread.sleep(200);
             log.info(Translator.translate("cloud.shutdown.started"));
-            boolean a = this.serverManager.stopAll();
-            if(a) log.info(Translator.translate("cloud.shutdown.servers"));
+            
+            // Only stop servers for STANDALONE or MASTER modes
+            if (!clusterConfig.enabled() || clusterConfig.mode() == ClusterMode.MASTER || clusterConfig.mode() == ClusterMode.STANDALONE) {
+                boolean a = this.serverManager.stopAll();
+                if(a) log.info(Translator.translate("cloud.shutdown.servers"));
+            }
+            
             Thread.sleep(500);
             this.pluginManager.disableAllPlugins();
             log.info(Translator.translate("cloud.shutdown.plugins"));
@@ -160,7 +181,11 @@ public class RedstoneCloud {
             log.info(Translator.translate("cloud.shutdown.complete"));
             this.scheduler.stopScheduler();
 
-            broker.getPool().getResource().flushDB();
+            // Only flush Redis DB in standalone mode or when we own the Redis instance
+            if (!clusterConfig.enabled() || clusterConfig.mode() == ClusterMode.STANDALONE) {
+                broker.getPool().getResource().flushDB();
+            }
+            
             broker.shutdown();
             if(redisInstance != null) redisInstance.shutdown();
         } catch (InterruptedException e) {
